LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/variations/at-font-face-font-matching-expected.txt

A A A A A A A1 A2 A2 A3 A3 A3A1 A2 A2 A3 A3 A3

PASS Descriptor matching priority: Stretch has higher priority than style
PASS Descriptor matching priority: Stretch has higher priority than weight
FAIL Descriptor matching priority: Style has higher priority than weight assert_equals: Unexpected font on test element expected 443 but got 488
PASS Matching font-weight: '400' should prefer '400' over '450 460'
PASS Matching font-weight: '400' should prefer '450 460' over '500'
PASS Matching font-weight: '400' should prefer '500' over '350 399'
FAIL Matching font-weight: '400' should prefer '350 399' over '351 398' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-weight: '400' should prefer '351 398' over '501 550'
PASS Matching font-weight: '400' should prefer '501 550' over '502 560'
PASS Matching font-weight: '430' should prefer '420 440' over '450 460'
PASS Matching font-weight: '430' should prefer '450 460' over '500'
PASS Matching font-weight: '430' should prefer '500' over '400 425'
PASS Matching font-weight: '430' should prefer '400 425' over '350 399'
PASS Matching font-weight: '430' should prefer '350 399' over '340 398'
PASS Matching font-weight: '430' should prefer '340 398' over '501 550'
FAIL Matching font-weight: '430' should prefer '501 550' over '502 560' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-weight: '500' should prefer '500' over '450 460'
FAIL Matching font-weight: '500' should prefer '450 460' over '400' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-weight: '500' should prefer '400' over '350 399'
PASS Matching font-weight: '500' should prefer '350 399' over '351 398'
PASS Matching font-weight: '500' should prefer '351 398' over '501 550'
FAIL Matching font-weight: '500' should prefer '501 550' over '502 560' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-weight: '501' should prefer '501' over '502 510'
FAIL Matching font-weight: '501' should prefer '502 510' over '503 520' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-weight: '501' should prefer '503 520' over '500'
FAIL Matching font-weight: '501' should prefer '500' over '450 460' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-weight: '501' should prefer '450 460' over '390 410'
PASS Matching font-weight: '501' should prefer '390 410' over '300 350'
PASS Matching font-weight: '399' should prefer '350 399' over '340 360'
PASS Matching font-weight: '399' should prefer '340 360' over '200 300'
FAIL Matching font-weight: '399' should prefer '200 300' over '400' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-weight: '399' should prefer '400' over '450 460'
PASS Matching font-weight: '399' should prefer '450 460' over '500 501'
PASS Matching font-weight: '399' should prefer '500 501' over '502 510'
FAIL Matching font-stretch: '100%' should prefer '100%' over '110% 120%' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-stretch: '100%' should prefer '110% 120%' over '115% 116%'
FAIL Matching font-stretch: '110%' should prefer '110% 120%' over '115% 116%' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-stretch: '110%' should prefer '115% 116%' over '105%'
FAIL Matching font-stretch: '110%' should prefer '105%' over '100%' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-stretch: '110%' should prefer '100%' over '50% 80%'
FAIL Matching font-stretch: '110%' should prefer '50% 80%' over '60% 70%' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-stretch: '90%' should prefer '90% 100%' over '50% 80%'
PASS Matching font-stretch: '90%' should prefer '50% 80%' over '60% 70%'
PASS Matching font-stretch: '90%' should prefer '60% 70%' over '110% 140%'
FAIL Matching font-stretch: '90%' should prefer '110% 140%' over '120% 130%' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'normal' should prefer 'normal' over 'oblique 0deg'
FAIL Matching font-style: 'normal' should prefer 'oblique 0deg' over 'oblique 10deg 40deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'normal' should prefer 'oblique 10deg 40deg' over 'oblique 20deg 30deg'
FAIL Matching font-style: 'normal' should prefer 'oblique 20deg 30deg' over 'oblique -50deg -20deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'normal' should prefer 'oblique -50deg -20deg' over 'oblique -40deg -30deg'
FAIL Matching font-style: 'italic' should prefer 'italic' over 'oblique 20deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'italic' should prefer 'oblique 20deg' over 'oblique 30deg 60deg'
FAIL Matching font-style: 'italic' should prefer 'oblique 30deg 60deg' over 'oblique 40deg 50deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'italic' should prefer 'oblique 40deg 50deg' over 'oblique 5deg 10deg'
FAIL Matching font-style: 'italic' should prefer 'oblique 5deg 10deg' over 'oblique 5deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'italic' should prefer 'oblique 5deg' over 'normal'
FAIL Matching font-style: 'italic' should prefer 'normal' over 'oblique 0deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'italic' should prefer 'oblique 0deg' over 'oblique -60deg -30deg'
FAIL Matching font-style: 'italic' should prefer 'oblique -60deg -30deg' over 'oblique -50deg -40deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'oblique 20deg' should prefer 'oblique 20deg' over 'oblique 30deg 60deg'
PASS Matching font-style: 'oblique 20deg' should prefer 'oblique 30deg 60deg' over 'oblique 40deg 50deg'
PASS Matching font-style: 'oblique 20deg' should prefer 'oblique 40deg 50deg' over 'oblique 10deg'
FAIL Matching font-style: 'oblique 20deg' should prefer 'oblique 10deg' over 'italic' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'oblique 20deg' should prefer 'italic' over 'oblique 0deg'
FAIL Matching font-style: 'oblique 20deg' should prefer 'oblique 0deg' over 'oblique -50deg -20deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'oblique 20deg' should prefer 'oblique -50deg -20deg' over 'oblique -40deg -30deg'
FAIL Matching font-style: 'oblique 21deg' should prefer 'oblique 21deg' over 'oblique 30deg 60deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'oblique 21deg' should prefer 'oblique 30deg 60deg' over 'oblique 40deg 50deg'
FAIL Matching font-style: 'oblique 21deg' should prefer 'oblique 40deg 50deg' over 'oblique 20deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'oblique 21deg' should prefer 'oblique 20deg' over 'oblique 10deg'
FAIL Matching font-style: 'oblique 21deg' should prefer 'oblique 10deg' over 'italic' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'oblique 21deg' should prefer 'italic' over 'oblique 0deg'
FAIL Matching font-style: 'oblique 21deg' should prefer 'oblique 0deg' over 'oblique -50deg -20deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'oblique 21deg' should prefer 'oblique -50deg -20deg' over 'oblique -40deg -30deg'
FAIL Matching font-style: 'oblique 10deg' should prefer 'oblique 10deg' over 'oblique 5deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'oblique 10deg' should prefer 'oblique 5deg' over 'oblique 15deg 20deg'
PASS Matching font-style: 'oblique 10deg' should prefer 'oblique 15deg 20deg' over 'oblique 30deg 60deg'
FAIL Matching font-style: 'oblique 10deg' should prefer 'oblique 30deg 60deg' over 'oblique 40deg 50deg' assert_equals: Unexpected font on test element expected 277 but got 533
FAIL Matching font-style: 'oblique 10deg' should prefer 'oblique 40deg 50deg' over 'italic' assert_equals: Unexpected font on test element expected 277 but got 533
FAIL Matching font-style: 'oblique 10deg' should prefer 'italic' over 'oblique 0deg' assert_equals: Unexpected font on test element expected 277 but got 533
FAIL Matching font-style: 'oblique 10deg' should prefer 'oblique 0deg' over 'oblique -50deg -20deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'oblique 10deg' should prefer 'oblique -50deg -20deg' over 'oblique -40deg -30deg'
PASS Matching font-style: 'oblique 0deg' should prefer 'oblique 0deg' over 'oblique 5deg'
PASS Matching font-style: 'oblique 0deg' should prefer 'oblique 5deg' over 'oblique 15deg 20deg'
PASS Matching font-style: 'oblique 0deg' should prefer 'oblique 15deg 20deg' over 'oblique 30deg 60deg'
FAIL Matching font-style: 'oblique 0deg' should prefer 'oblique 30deg 60deg' over 'oblique 40deg 50deg' assert_equals: Unexpected font on test element expected 277 but got 533
FAIL Matching font-style: 'oblique 0deg' should prefer 'oblique 40deg 50deg' over 'italic' assert_equals: Unexpected font on test element expected 277 but got 533
FAIL Matching font-style: 'oblique 0deg' should prefer 'italic' over 'oblique -50deg -20deg' assert_equals: Unexpected font on test element expected 277 but got 533
PASS Matching font-style: 'oblique 0deg' should prefer 'oblique -50deg -20deg' over 'oblique -40deg -30deg'
PASS Matching font-style: 'oblique -10deg' should prefer 'oblique -10deg' over 'oblique -5deg'
PASS Matching font-style: 'oblique -10deg' should prefer 'oblique -5deg' over 'oblique -1deg 0deg'
PASS Matching font-style: 'oblique -10deg' should prefer 'oblique -1deg 0deg' over 'oblique -20deg -15deg'
FAIL Matching font-style: 'oblique -10deg' should prefer 'oblique -20deg -15deg' over 'oblique -60deg -30deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'oblique -10deg' should prefer 'oblique -60deg -30deg' over 'oblique -50deg -40deg'
PASS Matching font-style: 'oblique -10deg' should prefer 'oblique -50deg -40deg' over 'italic'
FAIL Matching font-style: 'oblique -10deg' should prefer 'italic' over 'oblique 0deg 10deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'oblique -10deg' should prefer 'oblique 0deg 10deg' over 'oblique 40deg 50deg'
FAIL Matching font-style: 'oblique -20deg' should prefer 'oblique -20deg' over 'oblique -60deg -40deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'oblique -20deg' should prefer 'oblique -60deg -40deg' over 'oblique -10deg'
PASS Matching font-style: 'oblique -20deg' should prefer 'oblique -10deg' over 'italic'
FAIL Matching font-style: 'oblique -20deg' should prefer 'italic' over 'oblique 0deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'oblique -20deg' should prefer 'oblique 0deg' over 'oblique 30deg 60deg'
FAIL Matching font-style: 'oblique -20deg' should prefer 'oblique 30deg 60deg' over 'oblique 40deg 50deg' assert_equals: Unexpected font on test element expected 277 but got 533
PASS Matching font-style: 'oblique -21deg' should prefer 'oblique -21deg' over 'oblique -60deg -40deg'
PASS Matching font-style: 'oblique -21deg' should prefer 'oblique -60deg -40deg' over 'oblique -10deg'
PASS Matching font-style: 'oblique -21deg' should prefer 'oblique -10deg' over 'italic'
FAIL Matching font-style: 'oblique -21deg' should prefer 'italic' over 'oblique 0deg' assert_equals: Unexpected font on test element expected 488 but got 533
PASS Matching font-style: 'oblique -21deg' should prefer 'oblique 0deg' over 'oblique 30deg 60deg'
FAIL Matching font-style: 'oblique -21deg' should prefer 'oblique 30deg 60deg' over 'oblique 40deg 50deg' assert_equals: Unexpected font on test element expected 277 but got 533
